+++ cookiecutter/{{cookiecutter.organization}}-richie-site-factory/template/{{cookiecutter.site}}/src/backend/{{cookiecutter.site}}/settings.py
@@ -673,6 +673,13 @@ class Base(StyleguideMixin, DRFMixin, RichieCoursesConfigurationMixin, Configura
     # Define which node level can be processed to search for pageindex extension
     RICHIE_MAINMENUENTRY_MENU_ALLOWED_LEVEL = 0
 
+    # Whether you want to show the video iframe directly or prefer to lazy load it
+    RICHIE_VIDEO_PLUGIN_LAZY_LOADING = values.Value(
+        False,
+        environ_name="RICHIE_VIDEO_PLUGIN_LAZY_LOADING",
+        environ_prefix=None,
+    )
+
     # pylint: disable=invalid-name
     @property
     def ENVIRONMENT(self):